requirements.txt
groq
python-dotenv
rich
Pillow
pygame-ce
pygame_gui
